refactor(scheduler): extract snapshot system lifecycle to scheduler_snapshot_system.py
Phase C.1 of the scheduler decomposition extraction port.

Move the snapshot system lifecycle helpers — init_snapshot_system
(~330 lines including the post_forward_snapshot_callback closure) and
restore_snapshots_on_startup (~28 lines, previously nested inside the
"snapshot save/list/info/restore/eviction handlers" block) — out of
monolithic scheduler.py into a new fork-only file
python/sglang/srt/managers/scheduler_snapshot_system.py as module-level
functions taking a scheduler argument.

Per the §7 R5 re-scoping (committed at 90243d90), this extraction PR
uses module-level function relocation instead of attribute-ownership
migration. The Scheduler continues to own all ~10 snapshot/tier
attributes (snapshot_manager, snapshot_hook_manager, snapshot_policy,
host_pool, conversation_tracker, tier_manager, state_health_monitor,
_health_check_counter, pending_restore_registry, pending_restore_aliases).
The ~50+ cross-subsystem call sites elsewhere in scheduler.py are
UNCHANGED. External callers (test_startup_snapshot_restore.py) are
UNCHANGED because Scheduler keeps init_snapshot_system and
restore_snapshots_on_startup as 1-line delegators.

Module logger bound to "sglang.srt.managers.scheduler" so log records
continue to appear under the Scheduler logger name post-extraction.

Marker policy: header-only on the new file (Apache license + ENGRAM_MODIFIED
header, no internal BEGIN/END blocks) per snapshot.py precedent for
fully fork-authored files. The scheduler.py import + delegators are
wrapped in clean BEGIN/END ENGRAM blocks. The previously-nested
"restore_snapshots_on_startup implementation" marker inside S8's
wrapper is removed (block contents moved to the new file).

Semantics-preserving — no behavior change. Block balance 15/14 (the
pre-existing L29 orphan END / L30 unclosed BEGIN slated for Phase F).

py_compile passes on both touched files.
